scraping/requirements.txt
requests
beautifulsoup4
pillow
pystray
ollama
sqlite3  # Python ile birlikte gelir, ancak hata almamak için ekledim
logging  # Python'un standart kütüphanesi, eklemene gerek yok
